# optimization for parallelization when large core number