@vrabaud 2023-09-22
I confirm this is fine on our end.