@asmorkalov 2023-09-27
@opencv-alalek Could the patch be merged?